+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Added support for setting [PSR-18](https://www.php-fig.org/psr/psr-18/) HTTP client in
   `ClientRaw`.
 - Added support for [PSR-16](https://www.php-fig.org/psr/psr-16/) simple cache in `ClientRaw`.
+- Added method `User::getAvatarUrl()`.
 
 
 ## [1.1.0](https://github.com/badoo/jira-client/compare/v1.0.0...v1.1.0) (2020-02-04)